made requested changes
