+++ src/domain/discounts/new/index.tsx
@@ -61,16 +61,21 @@ const New: React.FC<NewProps> = ({ location }) => {
   }, [products])
 
   useEffect(() => {
-    if (discountType === "fixed" && selectedRegions.length > 1)
+    if (discountType === "fixed" && selectedRegions.length > 1) {
       setDiscountType("percentage")
+    }
   }, [selectedRegions, discountType])
 
   const methods = useForm()
 
   const getPersistedPrice = (price: number) => {
-    if (!selectedRegions.length || !regions) return
+    if (!selectedRegions.length || !regions) {
+      return
+    }
     const region = regions.find((r) => r.id === selectedRegions[0].value)
-    if (!region) return
+    if (!region) {
+      return
+    }
     return persistedPrice(region.currency_code, price)
   }
 
@@ -79,10 +84,11 @@ const New: React.FC<NewProps> = ({ location }) => {
       ...data,
       rule: {
         ...data.rule,
-        value:
-          discountType === "fixed"
-            ? getPersistedPrice(parseFloat(data.rule.value))
-            : parseFloat(data.rule.value),
+        value: isFreeShipping
+          ? 0
+          : discountType === "fixed"
+          ? getPersistedPrice(parseFloat(data.rule.value))
+          : parseFloat(data.rule.value),
         type: discountType,
         allocation: allocationItem ? "item" : "total",
         valid_for: appliesToAll
@@ -125,7 +131,9 @@ const New: React.FC<NewProps> = ({ location }) => {
   }
 
   useEffect(() => {
-    if (!toDuplicate) return
+    if (!toDuplicate) {
+      return
+    }
 
     hydrateDiscount({
       discount: toDuplicate,
